backend/requirement.txt
flask
flask-cors
requests
osmnx
scikit-learn
python-dotenv